Make parse functions collection-agnostic
This makes it possible to use a Set in one place and a Vec in another,
to log or count items without allocating a collection for all of them,
and to ignore errors with no overhead.

The alternative is converting them to custom iterators, which is more work
than I want to do right now.

Signed-off-by: Evan Lloyd New-Schmidt <evan@new-schmidt.com>